LPS-145917 Extract ccpp api and impl to modules.
