+++ modules/core/portal-bootstrap/system.packages.extra.bnd
@@ -314,7 +314,6 @@ Provide-Capability:\
 	@../../../lib/portal/asm-tree.jar,\
 	@../../../lib/portal/asm-util.jar,\
 	@../../../lib/portal/asm.jar,\
-	@../../../lib/portal/ccpp.jar,\
 	@../../../lib/portal/commons-beanutils.jar,\
 	@../../../lib/portal/commons-codec.jar,\
 	@../../../lib/portal/commons-collections.jar,\
