+++ portal-impl/src/com/liferay/portlet/internal/PortletRequestImpl.java
@@ -46,6 +46,7 @@ import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
+import com.liferay.portal.kernel.util.ServiceProxyFactory;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
@@ -178,7 +179,8 @@ public abstract class PortletRequestImpl implements LiferayPortletRequest {
 
 	public Object getCCPPProfile() {
 		if (_profile == null) {
-			_profile = PortalProfileFactory.getCCPPProfile(_httpServletRequest);
+			_profile = _portalProfileFactory.getCCPPProfile(
+				_httpServletRequest);
 		}
 
 		return _profile;
@@ -1261,6 +1263,11 @@ public abstract class PortletRequestImpl implements LiferayPortletRequest {
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		PortletRequestImpl.class);
+
+	private static volatile PortalProfileFactory _portalProfileFactory =
+		ServiceProxyFactory.newServiceTrackedInstance(
+			PortalProfileFactory.class, PortletRequestImpl.class,
+			"_portalProfileFactory", false);
 
 	private HttpServletRequest _httpServletRequest;
 	private boolean _invalidSession;
